+++ cmd/frontend/types/types.go
@@ -371,6 +371,23 @@ type SavedSearches struct {
 	OrgSavedSearches     int32
 }
 
+// Panel homepage represents interaction data on the
+// enterprise homepage panels.
+type HomepagePanels struct {
+	RecentFilesClickedPercentage           float64
+	RecentSearchClickedPercentage          float64
+	RecentRepositoriesClickedPercentage    float64
+	SavedSearchesClickedPercentage         float64
+	NewSavedSearchesClickedPercentage      float64
+	TotalPanelViews                        float64
+	UsersFilesClickedPercentage            float64
+	UsersSearchClickedPercentage           float64
+	UsersRepositoriesClickedPercentage     float64
+	UsersSavedSearchesClickedPercentage    float64
+	UsersNewSavedSearchesClickedPercentage float64
+	PercentUsersShown                      float64
+}
+
 // Secret represents the secrets table
 type Secret struct {
 	ID int32